+++ CHANGES.md
@@ -4,6 +4,10 @@ This files lists the changes during the lifetime of this project.
 
 ## unreleased
 
+## 2026-03-06
+
+- add endpoint to worker container for health checks
+
 ## 2026-03-05
 
 - 189: Add multi-select checkbox filters